Add new props to App component
Two new props were added to the App component: one that accepts
the name of the response header that stores the CSRF token and
one that accepts the name of the field in the body of the response
that contains the CSRF token. Only one of these two props must be
assigned a value and responses from the server do not need to
contain both, that would be redundant.