Refactor mock server definitions
The mock server request handlers used in the end-to-end tests
were refactored. They now allow the tester to specify the name
of the header or the field name to use when sending the CSRF
token in requests.

Furthermore, one of the tests previously didn't actually test for
anything. It has been updated and now ensures that the contact
form always gives preference to the header method for sending
the CSRF token if both a header name and field name are specified.