Ensure POST request uses CSRF field name
The POST request that is sent when a user submits the contact
form previously didn't use the CSRF field name specified in the
csrfFieldName variable. Instead, it used a hard-coded value of
"token" as the field name in the POST request body. This has
been fixed, and now all POST requests will use the value of
csrfFieldName as the name of the CSRF token value in the body of
the request.